Refactor ProjectEditor js to use a Store
Still not super Flux-y because I am not using Actions from a Dispatcher,
but I am pretty happy to have all the XHR-related stuff in one place, and
only one Component knows about the Store.

Not happy with the method of declaring the Store on window, but react-rails'
react_component method doesn't allow wiring in external JS objects.