Added draft status to projects
Adds draft status to projects and filters project index to only display
published projects. Existing projects are set to published and new
projects are set to draft by default. Trying to view a challenge that
isn't published if you aren't an admin will redirect you to the
root_path currently.